Fix for Windows x86_64; size of SCardContext should be Pointer.SIZE.
Previously, I assumed that SCardContext was always a NativeLong. This is correct on Linux, but wrong on Windows and Mac. All functions in Windows x86_64 would fail with error 6 (ERROR_INVALID_HANDLE) because I truncated the pointer. I did not notice any crashes on Mac, because it SCARDCONTEXT there is just a int32 that increments from 1.

Fix SCardContext to match the different headers: make it a subclass of IntegerType instead of Structure, and make it 32 bits on Mac, the same size as a pointer on Windows, and the same size as a long on Linux.