Simplify Winscard by replacing int or NativeLong with Dword type.
Add Dword and DwordByReference types. With the Dword type that is always 32 bits on Windows and Mac but long on Linux, we can unify the function and struct declarations that we give to JNA.

Use Dword in SCardReaderState Structure. Delete WinscardSCardReaderState, OSXSCardReaderState, LinuxSCardReaderState.

Use Dword in WinscardLibrary. Delete PcscLiteLibrary and PcscLiteAdapter which are now redundant.